requirements.txt
asgiref==3.2.7
Django==3.0.5
django-colorfield==0.3.0
django-polls==0.1
Pillow==7.1.1
mysqlclient==1.4.6
pytz==2019.2
sqlparse==0.3.0
python-dotenv
